Reduction of CUDA compiler warnings.

+++ mundy/mesh/tests/performance_tests/PerfTestAccessors.cpp
@@ -128,20 +128,20 @@ void run_test() {
     for (size_t i = 0; i < num_nodes; ++i) {
       stk::mesh::Entity node = nodes[i];
       double& scalar = stk::mesh::field_data(scalar_field, node)[0];
-      scalar += i;
+      scalar += static_cast<double>(i);
     }
   }
-  double scalar_field_time = scalar_field_timer.seconds() / num_trials;
+  double scalar_field_time = scalar_field_timer.seconds() / static_cast<double>(num_trials);
 
   Kokkos::Timer scalar_accessor_timer;
   for (size_t t = 0; t < num_trials; ++t) {
     for (size_t i = 0; i < num_nodes; ++i) {
       stk::mesh::Entity node = nodes[i];
       auto scalar = scalar_accessor(node);
-      scalar[0] += i;
+      scalar[0] += static_cast<double>(i);
     }
   }
-  double scalar_accessor_time = scalar_accessor_timer.seconds() / num_trials;
+  double scalar_accessor_time = scalar_accessor_timer.seconds() / static_cast<double>(num_trials);
   std::cout << "Scalar field time: " << scalar_field_time << " vs Scalar accessor time: " << scalar_accessor_time
             << std::endl;
   std::cout << " Ratio (accessor/field): " << scalar_accessor_time / scalar_field_time << std::endl;
@@ -155,11 +155,11 @@ void run_test() {
       stk::mesh::Entity node = nodes[i];
       for (size_t j = 0; j < 3; ++j) {
         double& vector3 = stk::mesh::field_data(vector3_field, node)[j];
-        vector3 += i * j;
+        vector3 += static_cast<double>(i) * static_cast<double>(j);
       }
     }
   }
-  double vector3_field_time = vector3_field_timer.seconds() / num_trials;
+  double vector3_field_time = vector3_field_timer.seconds() / static_cast<double>(num_trials);
 
   Kokkos::Timer vector3_accessor_timer;
   for (size_t t = 0; t < num_trials; ++t) {
@@ -167,11 +167,11 @@ void run_test() {
       stk::mesh::Entity node = nodes[i];
       auto vector3 = vector3_accessor(node);
       for (size_t j = 0; j < 3; ++j) {
-        vector3[j] += i * j;
+        vector3[j] += static_cast<double>(i) * static_cast<double>(j);
       }
     }
   }
-  double vector3_accessor_time = vector3_accessor_timer.seconds() / num_trials;
+  double vector3_accessor_time = vector3_accessor_timer.seconds() / static_cast<double>(num_trials);
   std::cout << "Vector3 field time: " << vector3_field_time << " vs Vector3 accessor time: " << vector3_accessor_time
             << std::endl;
   std::cout << " Ratio (accessor/field): " << vector3_accessor_time / vector3_field_time << std::endl;
@@ -184,11 +184,11 @@ void run_test() {
       stk::mesh::Entity node = nodes[i];
       for (size_t j = 0; j < 9; ++j) {
         double& matrix3 = stk::mesh::field_data(matrix3_field, node)[j];
-        matrix3 += i * j;
+        matrix3 += static_cast<double>(i) * static_cast<double>(j);
       }
     }
   }
-  double matrix3_field_time = matrix3_field_timer.seconds() / num_trials;
+  double matrix3_field_time = matrix3_field_timer.seconds() / static_cast<double>(num_trials);
 
   Kokkos::Timer matrix3_accessor_timer;
   for (size_t t = 0; t < num_trials; ++t) {
@@ -196,11 +196,11 @@ void run_test() {
       stk::mesh::Entity node = nodes[i];
       auto matrix3 = matrix3_accessor(node);
       for (size_t j = 0; j < 9; ++j) {
-        matrix3[j] += i * j;
+        matrix3[j] += static_cast<double>(i) * static_cast<double>(j);
       }
     }
   }
-  double matrix3_accessor_time = matrix3_accessor_timer.seconds() / num_trials;
+  double matrix3_accessor_time = matrix3_accessor_timer.seconds() / static_cast<double>(num_trials);
   std::cout << "Matrix3 field time: " << matrix3_field_time << " vs Matrix3 accessor time: " << matrix3_accessor_time
             << std::endl;
   std::cout << " Ratio (accessor/field): " << matrix3_accessor_time / matrix3_field_time << std::endl;
@@ -214,11 +214,11 @@ void run_test() {
       stk::mesh::Entity node = nodes[i];
       for (size_t j = 0; j < 4; ++j) {
         double& quaternion = stk::mesh::field_data(quaternion_field, node)[j];
-        quaternion += i * j;
+        quaternion += static_cast<double>(i) * static_cast<double>(j);
       }
     }
   }
-  double quaternion_field_time = quaternion_field_timer.seconds() / num_trials;
+  double quaternion_field_time = quaternion_field_timer.seconds() / static_cast<double>(num_trials);
 
   Kokkos::Timer quaternion_accessor_timer;
   for (size_t t = 0; t < num_trials; ++t) {
@@ -226,11 +226,11 @@ void run_test() {
       stk::mesh::Entity node = nodes[i];
       auto quaternion = quaternion_accessor(node);
       for (size_t j = 0; j < 4; ++j) {
-        quaternion[j] += i * j;
+        quaternion[j] += static_cast<double>(i) * static_cast<double>(j);
       }
     }
   }
-  double quaternion_accessor_time = quaternion_accessor_timer.seconds() / num_trials;
+  double quaternion_accessor_time = quaternion_accessor_timer.seconds() / static_cast<double>(num_trials);
   std::cout << "Quaternion field time: " << quaternion_field_time
             << " vs Quaternion accessor time: " << quaternion_accessor_time << std::endl;
   std::cout << " Ratio (accessor/field): " << quaternion_accessor_time / quaternion_field_time << std::endl;
@@ -244,11 +244,11 @@ void run_test() {
       stk::mesh::Entity node = nodes[i];
       for (size_t j = 0; j < 6; ++j) {
         double& aabb = stk::mesh::field_data(aabb_field, node)[j];
-        aabb += i * j;
+        aabb += static_cast<double>(i) * static_cast<double>(j);
       }
     }
   }
-  double aabb_field_time = aabb_field_timer.seconds() / num_trials;
+  double aabb_field_time = aabb_field_timer.seconds() / static_cast<double>(num_trials);
 
   Kokkos::Timer aabb_accessor_timer;
   for (size_t t = 0; t < num_trials; ++t) {
@@ -256,11 +256,11 @@ void run_test() {
       stk::mesh::Entity node = nodes[i];
       auto aabb = aabb_accessor(node);
       for (size_t j = 0; j < 6; ++j) {
-        aabb[j] += i * j;
+        aabb[j] += static_cast<double>(i) * static_cast<double>(j);
       }
     }
   }
-  double aabb_accessor_time = aabb_accessor_timer.seconds() / num_trials;
+  double aabb_accessor_time = aabb_accessor_timer.seconds() / static_cast<double>(num_trials);
   std::cout << "AABB field time: " << aabb_field_time << " vs AABB accessor time: " << aabb_accessor_time << std::endl;
   std::cout << " Ratio (accessor/field): " << aabb_accessor_time / aabb_field_time << std::endl;
 }
